Created the "fidocadj.dialogs.controls" package to encapsulate all auxiliary classes for dialogs
