+++ src/fidocadj/dialogs/DialogLayer.java
@@ -1,5 +1,6 @@
 package fidocadj.dialogs;
 
+import fidocadj.dialogs.controls.LayerCellRenderer;
 import java.io.*;
 import java.awt.*;
 import java.awt.event.*;
