+++ src/fidocadj/dialogs/controls/CellDash.java
@@ -1,4 +1,4 @@
-package fidocadj.dialogs;
+package fidocadj.dialogs.controls;
 
 import java.io.*;
 import java.awt.*;
