+++ src/fidocadj/dialogs/controls/EnterCircuitFrame.java
@@ -1,4 +1,5 @@
-package fidocadj.dialogs;
+package fidocadj.dialogs.controls;
+import fidocadj.dialogs.DialogUtil;
 import java.awt.*;
 
 import javax.swing.*;
