+++ src/fidocadj/dialogs/controls/LayerInfo.java
@@ -1,11 +1,8 @@
-package fidocadj.dialogs;
+package fidocadj.dialogs.controls;
 
-/** This class contains some settings about the actual dashing style. It is used
-    in the automatic primitive characteristics dialog.
+/**
 
-    @author Davide Bucci
-
-    <pre>
+<pre>
     This file is part of FidoCadJ.
 
     FidoCadJ is free software: you can redistribute it and/or modify
@@ -23,23 +20,20 @@ package fidocadj.dialogs;
     @see <a href=http://www.gnu.org/licenses/>http://www.gnu.org/licenses/</a>.
 
     Copyright 2009-2023 by Davide Bucci
-    </pre>
-
+</pre>
+    @author Davide Bucci
 */
-public class DashInfo
+public class LayerInfo
 {
-    // Here we store the dash style
-    public int style;
+    int layer;
 
-    /** Creator.
-        @param i the style to be stored
-    */
-    public DashInfo(int i)
-    { style=i; }
+    /** Create a LayerInfo object with the given layer
+    @param i the layer to be used.
 
-    /** Retrieve the style.
-        @return the style.
     */
-    public int getStyle()
-    { return style; }
+    public LayerInfo(int i)
+    { layer=i; }
+
+    public int getLayer()
+    { return layer; }
 }
