+++ src/fidocadj/dialogs/controls/OriginCircuitPanel.java
@@ -1,4 +1,4 @@
-package fidocadj.dialogs;
+package fidocadj.dialogs.controls;
 
 
 import fidocadj.circuit.CircuitPanel;
