+++ src/fidocadj/dialogs/controls/package-info.java
@@ -1,0 +1,5 @@
+/**
+    Ancillary classes for dialogs of FidoCadJ.
+*/
+
+package fidocadj.dialogs.controls;
