+++ src/fidocadj/primitives/PrimitiveAdvText.java
@@ -3,8 +3,8 @@ package fidocadj.primitives;
 import java.io.*;
 import java.util.*;
 
-import fidocadj.dialogs.ParameterDescription;
-import fidocadj.dialogs.LayerInfo;
+import fidocadj.dialogs.controls.ParameterDescription;
+import fidocadj.dialogs.controls.LayerInfo;
 import fidocadj.export.ExportInterface;
 import fidocadj.geom.MapCoordinates;
 import fidocadj.geom.GeometricDistances;
